add parseTemplateString
