+++ ui/src/pages/namespace/Explorer/Page/poc/PageCompiler/primitives/TemplateString.tsx
@@ -1,0 +1,15 @@
+import { TemplateStringType } from "../../schema/primitives/templateString";
+import { Variable } from "./Variable";
+import { parseTemplateString } from "./Variable/utils";
+
+type TemplateStringProps = {
+  value: TemplateStringType;
+};
+
+export const TemplateString = ({ value }: TemplateStringProps) => {
+  const templateFragments = parseTemplateString(value, (match, index) => (
+    <Variable key={index} value={match} />
+  ));
+
+  return templateFragments;
+};
